+++ examples/serve/src/fixtures/audience/__pointCutConfig.js
@@ -1,11 +1,14 @@
 import { basename, posix } from "path";
-import getToggleHandlerPath from "../../getToggleHandlerPath.js";
+import loadStrategyFactory from "@asos/web-toggle-point-webpack/moduleLoadStrategyFactories/deferredDynamicImportLoadStrategyFactory";
 
 export default {
   name: "audience",
-  togglePointModule: "/src/fixtures/audience/__togglePoint.js",
+  togglePointModuleSpecifier: "/src/fixtures/audience/__togglePoint.js",
   variantGlob: "./src/fixtures/audience/**/cohort-[1-9]*([0-9])/*.js",
-  toggleHandler: getToggleHandlerPath("singlePathSegment.js"),
+  toggleHandlerFactoryModuleSpecifier: import.meta.resolve(
+    "../../toggleHandlerFactories/singlePathSegment.js"
+  ),
   joinPointResolver: (path) =>
-    posix.resolve(path, "../..", basename(path).replace("bespoke", "control"))
+    posix.resolve(path, "../..", basename(path).replace("bespoke", "control")),
+  loadStrategy: loadStrategyFactory()
 };