+++ examples/serve/src/fixtures/config/__pointCutConfig.js
@@ -1,10 +1,11 @@
-import getToggleHandlerPath from "../../getToggleHandlerPath.js";
 import joinPointResolver from "../../joinPointResolver.js";
 
 export default {
   name: "configuration",
-  togglePointModule: "/src/fixtures/config/__togglePoint.js",
+  togglePointModuleSpecifier: "/src/fixtures/config/__togglePoint.js",
   variantGlob: "./src/fixtures/config/**/sites/*/*.js",
-  toggleHandler: getToggleHandlerPath("listExtractionFromPathSegment.js"),
+  toggleHandlerFactoryModuleSpecifier: import.meta.resolve(
+    "../../toggleHandlerFactories/listExtractionFromPathSegment.js"
+  ),
   joinPointResolver
 };