+++ examples/serve/src/fixtures/config/__togglePoint.js
@@ -1,9 +1,9 @@
 import featuresStore from "./__featuresStore.js";
 
-export default (joinPoint, featuresMap) => {
+export default ({ joinPoint, featuresMap, unpack }) => {
   const site = featuresStore.getFeatures();
   if (featuresMap.has(site)) {
-    return featuresMap.get(site).default;
+    return unpack(featuresMap.get(site)).default;
   }
-  return joinPoint.default;
+  return unpack(joinPoint).default;
 };