+++ examples/serve/src/fixtures/event/__togglePoint.js
@@ -1,9 +1,9 @@
 import featuresStore from "./__featuresStore.js";
 
-export default (joinPoint, featuresMap) => {
+export default ({ joinPoint, featuresMap, unpack }) => {
   const event = featuresStore.getFeatures();
   if (featuresMap.has(event)) {
-    return featuresMap.get(event).default;
+    return unpack(featuresMap.get(event)).default;
   }
-  return joinPoint;
+  return unpack(joinPoint).default;
 };