+++ examples/serve/src/fixtures/translation/__pointCutConfig.js
@@ -1,8 +1,10 @@
 import joinPointResolver from "../../joinPointResolver.js";
+import loadStrategyFactory from "@asos/web-toggle-point-webpack/moduleLoadStrategyFactories/staticLoadStrategyFactory";
 
 export default {
   name: "translation",
-  togglePointModule: "/src/fixtures/translation/__togglePoint.js",
+  togglePointModuleSpecifier: "/src/fixtures/translation/__togglePoint.js",
   variantGlob: "./src/fixtures/translation/languages/*/*.json",
-  joinPointResolver
+  joinPointResolver,
+  loadStrategy: loadStrategyFactory()
 };