+++ examples/serve/src/getToggleHandlerPath.js
@@ -1,6 +1,0 @@
-import { posix } from "path";
-
-const getToggleHandlerPath = (...paths) =>
-  posix.join("/", "src", "toggleHandlers", ...paths);
-
-export default getToggleHandlerPath;