+++ examples/serve/docs/CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - ??
+
+### Changed
+
+- moved to `production` webpack mode with `source-map` devtool and `source-map-loader`, for clarity when using dev tools
+- updates for new `webpack` package:
+  - moving to `toggleHandlerFactories` from `toggleHandlers`
+  - using non-default `loadStrategy` for two of the four examples
+    - `audience` becomes `deferredImport` (`async`)
+    - `translation` becomes `static` (which was the previous default...)
+    - other default to `deferredRequire` (synchronous, but delayed require of module)
+- move to use `import.meta.resolve` replacing the hand-rolled `getToggleHandlerPath.js`
+
 ## [0.2.3] - 2025-02-27
 
 ### Changed